+++ .coverage
@@ -1,1 +1,1 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\project\\dec_levelup_andela\\app\\__init__.py":[1,3,5,6,7,8,9],"C:\\project\\dec_levelup_andela\\app\\routes\\__init__.py":[1],"C:\\project\\dec_levelup_andela\\app\\routes\\get_all_route.py":[1,2,4,6,7],"C:\\project\\dec_levelup_andela\\app\\views\\__init__.py":[1,2,4,5,7,8,9],"C:\\project\\dec_levelup_andela\\app\\models\\__init__.py":[1],"C:\\project\\dec_levelup_andela\\app\\models\\corrupt_model.py":[1,2,4,7,20,35],"C:\\project\\dec_levelup_andela\\app\\models\\usr_model.py":[1,3,6,18],"C:\\project\\dec_levelup_andela\\app\\views\\get_all_view.py":[1,3,4,5,6,7],"C:\\project\\dec_levelup_andela\\app\\routes\\create_redflag_route.py":[1,2,4,6,7],"C:\\project\\dec_levelup_andela\\app\\views\\create_redflag.py":[1,6,8],"C:\\project\\dec_levelup_andela\\app\\routes\\create_user_route.py":[1,2,4,5,6],"C:\\project\\dec_levelup_andela\\app\\views\\create_new_user.py":[1,5,7],"C:\\project\\dec_levelup_andela\\app\\routes\\get_one_route.py":[1,2,4,6,7],"C:\\project\\dec_levelup_andela\\app\\routes\\delete_route.py":[1,2,4,6,7],"C:\\project\\dec_levelup_andela\\app\\tests\\__init__.py":[1],"C:\\project\\dec_levelup_andela\\app\\tests\\test_base.py":[1,2,3,4,9,10,16,11,12],"C:\\project\\dec_levelup_andela\\app\\tests\\test_create.py":[1,2,7,9,10,11,12],"C:\\project\\dec_levelup_andela\\app\\tests\\test_create_new_user.py":[1,2,4,6,7,8],"C:\\project\\dec_levelup_andela\\app\\tests\\test_delete.py":[1,2,7,9,10,11],"C:\\project\\dec_levelup_andela\\app\\tests\\test_get_all_route.py":[1,2,7,9,10,11,12],"C:\\project\\dec_levelup_andela\\app\\tests\\test_get_one_redflag.py":[1,2,4,6,7,8],"C:\\project\\dec_levelup_andela\\app\\views\\delete_red_flag.py":[],"C:\\project\\dec_levelup_andela\\app\\views\\get_one_view.py":[]}}
+!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\project\\dec_levelup_andela\\app\\__init__.py":[1,3,5,6,7,8,9],"C:\\project\\dec_levelup_andela\\app\\routes\\__init__.py":[1],"C:\\project\\dec_levelup_andela\\app\\routes\\get_all_route.py":[1,2,4,6,7],"C:\\project\\dec_levelup_andela\\app\\views\\__init__.py":[1,2,4,5,7,8,9],"C:\\project\\dec_levelup_andela\\app\\models\\__init__.py":[1],"C:\\project\\dec_levelup_andela\\app\\models\\corrupt_model.py":[1,2,4,7,20,35],"C:\\project\\dec_levelup_andela\\app\\models\\usr_model.py":[1,3,6,18],"C:\\project\\dec_levelup_andela\\app\\views\\get_all_view.py":[1,4,5],"C:\\project\\dec_levelup_andela\\app\\routes\\create_redflag_route.py":[1,2,4,6,7],"C:\\project\\dec_levelup_andela\\app\\views\\create_redflag.py":[1,6,8],"C:\\project\\dec_levelup_andela\\app\\routes\\create_user_route.py":[1,2,4,5,6],"C:\\project\\dec_levelup_andela\\app\\views\\create_new_user.py":[1,5,7],"C:\\project\\dec_levelup_andela\\app\\routes\\get_one_route.py":[1,2,4,6,7],"C:\\project\\dec_levelup_andela\\app\\routes\\delete_route.py":[1,2,4,6,7],"C:\\project\\dec_levelup_andela\\app\\tests\\__init__.py":[1],"C:\\project\\dec_levelup_andela\\app\\tests\\test_base.py":[1,2,3,4,9,10,16,11,12],"C:\\project\\dec_levelup_andela\\app\\tests\\test_incident.py":[1,2,4,6,7,8],"C:\\project\\dec_levelup_andela\\app\\tests\\test_user.py":[1,3,5,6,7],"C:\\project\\dec_levelup_andela\\app\\views\\delete_red_flag.py":[],"C:\\project\\dec_levelup_andela\\app\\views\\get_one_view.py":[]}}
